refactor: update imports for Gemini feature consolidation
- Created new barrel files for feature exports
- Updated import paths in key components
- Simplified imports using @/features/gemini paths
- Added central feature index.ts for public API
