+++ package.json
@@ -97,17 +97,23 @@
     "concurrently": "^8.0.0",
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
+    "dependency-cruiser": "^16.10.0",
     "eslint": "^9.9.0",
+    "eslint-plugin-complexity": "^1.0.2",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.9",
+    "eslint-plugin-sonarjs": "^3.0.2",
     "express": "^4.18.2",
     "globals": "^15.9.0",
+    "jscpd": "^4.0.5",
     "jsdom": "^26.0.0",
+    "madge": "^8.0.0",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.11",
     "typescript": "^5.5.3",
     "typescript-eslint": "^8.0.1",
     "vite": "^6.2.3",
-    "vitest": "^3.0.9"
+    "vitest": "^3.0.9",
+    "webpack-bundle-analyzer": "^4.10.2"
   }
 }
