+++ src/components/ui/ai-chat/ChatMessageList.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import { Bot } from "lucide-react";
-import { AIMessage } from "@/services/chat/messageTypes";
+import { AIMessage } from "@/features/gemini/types/messageTypes";
 import { TypingIndicator } from "./TypingIndicator";
 
 interface ChatMessageProps {
