+++ src/features/gemini/index.ts
@@ -1,4 +1,15 @@
-export * from "./hooks";
-export * from "./services";
-export * from "./types";
-export * from "./config/GeminiConfig";
+// Core exports
+export * from './services/geminiAdapter';
+export * from './services/initialize';
+export * from './services/formatters';
+
+// Config exports
+export * from './config/GeminiConfig';
+
+// Hook exports
+export * from './hooks/useGeminiMessageSubmission';
+export * from './hooks/useGeminiInitialization';
+export * from './hooks/useGeminiAudio';
+
+// Type exports
+export * from './types';
