+++ src/hooks/useGeminiService.ts
@@ -1,3 +1,10 @@
+import { 
+  GeminiAdapter,
+  GeminiConfig,
+  useGeminiMessageSubmission,
+  useGeminiInitialization,
+  useGeminiAudio,
+} from '@/features/gemini';
 // If this file exists, it likely uses the data
 
 import { useState, useCallback } from 'react';
